Organise les scripts de la scène Commencement
