+++ Assets/Scripts/MonoBehavioursUsed/InstantiateGameObject.cs
@@ -1,18 +1,30 @@
 using UnityEngine;
 
+/// <summary>
+/// Utilitaire simple permettant d'instancier un prefab soit au démarrage,
+/// soit sur demande depuis une Timeline.
+/// </summary>
 public class InstantiateGameObject : MonoBehaviour
 {
+    [Tooltip("Si vrai, le prefab est instancié dès Start().")]
     public bool instantiateAtStart = false;
+
+    [Tooltip("Prefab à instancier.")]
     public GameObject gameObjectToInstantiate;
 
     void Start()
     {
+        // Instancie automatiquement l'objet au lancement si l'option est activée.
         if (instantiateAtStart)
         {
             Instantiate(gameObjectToInstantiate, transform.position, Quaternion.identity);
         }
     }
 
+    /// <summary>
+    /// Instancie un GameObject depuis une timeline ou un autre système externe.
+    /// </summary>
+    /// <param name="gameObjectToInstantiate">Prefab à instancier.</param>
     public void InstantiateFromTimeline(GameObject gameObjectToInstantiate)
     {
         // Vérifie les références avant d'instancier pour éviter les NullReferenceException
